Okay I actually need the services to talk to the database, so I am implementing api endpoints.
The pipeline to the database should be mostly constructed. and what I am looking at next is a create user screen.

I did not anticipate so much of my time going into login. but thems the breaks. It is possible that Instead of diving into the rabbit hole of how to create users in a secure fashion. that I just have an admin user in the database that I login as. and move on to having a react front end that starts displaying to me kafka streams in various ways.