register user endpoint created, checks to see if the user is there, and if not, creates a JWT token and authenticate the user
fix for front end referenceing the correct property in the response object,

the generated token actually never had roles attached to it until now

fetching the user information in profile does not work because it is kept in a separate database that is not implemented yet.

For that I want to implement kafka, where the user details service publishes a message that the user service reads from, and then creates the user information with.